chore: move placeholder logic to SDK
