+++ discord/src/main/kotlin/gg/airbrush/discord/events/PlayerJoin.kt
@@ -15,8 +15,8 @@ package gg.airbrush.discord.events
 import gg.airbrush.discord.bot
 import gg.airbrush.discord.discordConfig
 import gg.airbrush.discord.eventNode
-import gg.airbrush.discord.lib.Placeholder
-import gg.airbrush.discord.lib.parsePlaceholders
+import gg.airbrush.sdk.lib.Placeholder
+import gg.airbrush.sdk.lib.parsePlaceholders
 import net.minestom.server.entity.Player
 import net.minestom.server.event.player.PlayerDisconnectEvent
 import net.minestom.server.event.player.PlayerSpawnEvent
