+++ sdk/src/main/kotlin/gg/airbrush/sdk/lib/Placeholder.kt
@@ -1,9 +1,7 @@
-
-
 /*
  * This file is part of Airbrush
  *
- * Copyright (c) 2023 Airbrush Team
+ * Copyright (c) Airbrush Team
  *
  * This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
  *
@@ -12,7 +10,7 @@
  * You should have received a copy of the GNU General Public License along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package gg.airbrush.discord.lib
+package gg.airbrush.sdk.lib
 
 class Placeholder(val string: String, val replacement: String)
